Refactor in components
